% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_data_generator_function.R
\name{generate_sample_meta_data}
\alias{generate_sample_meta_data}
\title{generate_sample_meta_data
Function to create the sample meta data using input from the collaborator}
\usage{
generate_sample_meta_data(
  project_id = vectorize_input("Enter Project ID: "),
  conditions = vectorize_input("Enter conditions (Comma Separated): ", ","),
  matrices = vectorize_input("Enter matrices (Space Separated): "),
  replicates = vectorize_input("Enter number of replicates: ", convert = as.numeric),
  num_subjects = vectorize_input("Enter number of subjects in the study: ", convert =
    as.numeric),
  path = "Sample_Meta_Data.csv"
)
}
\arguments{
\item{project_id}{A unique identifier for the project}

\item{conditions}{A vector of conditions in the study}

\item{matrices}{A vector of sample types used in the study}

\item{replicates}{The number of replicates for each subject}

\item{num_subjects}{The number of subjects in the study}

\item{path}{The path to save the csv file (set to NULL to skip save)}
}
\value{
A data.frame with the sample meta data is invisibly returned and a csv file is generated at `path`.
}
\description{
generate_sample_meta_data
Function to create the sample meta data using input from the collaborator
}
\examples{
generate_sample_meta_data(project_id = 'abc123',
                          conditions = c('stim', 'control'),
                          matrices = c('plasma', 'kidney', 'heart'),
                          replicates = 3,
                          num_subjects = 5,
                          path = NULL)
}
