% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence.R
\name{generate_sequence}
\alias{generate_sequence}
\title{Generate SciexOS Formatted Sequence}
\usage{
generate_sequence(
  plate_loading,
  qc_plate_data,
  blank_plate_data,
  project_id,
  injection_vol
)
}
\arguments{
\item{plate_loading}{Data frame containing autosampler plate loading information.}

\item{qc_plate_data}{Data frame containing QC metadata.}

\item{blank_plate_data}{Data frame containing BLANK metadata.}

\item{project_id}{Unique project identifier.}

\item{injection_vol}{Numeric value specifying the injection volume.}
}
\value{
Exports sequence CSV files for SciexOS import.
}
\description{
This function generates a randomized mass spec sequence, inserting QC and BLANK samples at specified intervals.
}
