% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_acquisition_ids_reduced.R
\name{process_acquisition_ids_reduced}
\alias{process_acquisition_ids_reduced}
\title{Process TAS File in preparation for plate meta data generation}
\usage{
process_acquisition_ids_reduced(pre_processed_data)
}
\arguments{
\item{submitted_sample_data}{A data frame containing acquisition IDs, with column `Submitted Sample Ids`.}
}
\value{
A data frame with expanded sample submission data, including generated acquisition IDs.
}
\description{
This function parses the file name to incorporate meta data with the batch information.
Creates a unique project ID that is required downstream in the sequence generation pipeline.
}
\details{
This is a largely reduced version from the original.
Will be maintained in case the acquisition ID generation strategy continues.
}
