% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_data_generator_function.R
\name{generate_plate_meta_data}
\alias{generate_plate_meta_data}
\title{generate_plate_meta_data
Function to create the plate meta data using input from the collaborator}
\usage{
generate_plate_meta_data(
  sample_meta_data,
  num_wells = 54,
  num_blanks = 1,
  num_qc = 1,
  plates_per_batch = 3,
  path = NULL,
  randomize = FALSE,
  seed = 394875
)
}
\arguments{
\item{sample_meta_data}{A data.frame with the sample meta data (see Details)}

\item{num_wells}{Number of wells per plate}

\item{num_blanks}{Number of blank wells per batch}

\item{num_qc}{Number of QC wells per matrix type}

\item{plates_per_batch}{Number of plates to use per batch}

\item{path}{The path to save the csv file (set to NULL to skip save)}

\item{randomize}{Logical, whether to randomize the sample order if they need to be run in multiple batches}

\item{seed}{Integer, the seed to use for randomization}
}
\value{
A data.frame with the plate meta data is invisibly returned and a csv file is generated at `path`.
}
\description{
generate_plate_meta_data
Function to create the plate meta data using input from the collaborator
}
\details{
The `sample_meta_data` data.frame should have the following columns: `Project_id`, `Submitted Sample Ids`, and `Matrix`.
Other columns are allowed but will be ignored.
}
