% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_acquisition_ids.R,
%   R/process_acquisition_ids_reduced.R
\name{process_acquisition_ids}
\alias{process_acquisition_ids}
\title{Process sample submission data and generate acquisition IDs}
\usage{
process_acquisition_ids(pre_processed_data, selected_ms_methods)

process_acquisition_ids(pre_processed_data, selected_ms_methods)
}
\arguments{
\item{selected_ms_methods}{A character vector of selected MS methods.}

\item{submitted_sample_data}{A data frame containing acquisition IDs, with column `Submitted Sample Ids`.}
}
\value{
A data frame with expanded sample submission data, including generated acquisition IDs.

A data frame with expanded sample submission data, including generated acquisition IDs.
}
\description{
This function processes TAS sample submission data by generating unique acquisition IDs
based on the file name structure and selected mass spectrometry (MS) methods.

This function parses the file name to incorporate meta data with the batch information.
Creates a unique project ID that is required downstream in the sequence generation pipeline.
}
\details{
This is a largely reduced version from the original.
Will be maintained in case the acquisition ID generation strategy continues.
}
